006/12 11:45

+++ hello/src/main/java/soccer/hello/Repository/PlayerRepository.java
@@ -18,7 +18,7 @@ public interface PlayerRepository {
 
     public List<Player> getAllPlayers();
 
-    public Optional<Player> findPlayer(int playerId);
+    public Optional<Player> findPlayer(Long playerId);
 
 
 }
