+++ hello/src/main/java/soccer/hello/service/PlayerServi.java
@@ -10,7 +10,7 @@ public interface PlayerServi {
 
     public List<Player> getAllPlayers();
 
-    public Optional<Player> findPlayer(int playerId);
+    public Optional<Player> findPlayer(Long playerId);
